requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.21.4
pandas==1.3.4
scikit-learn==1.0.1

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage==1.43.0
mlflow
s3fs
fastapi
uvicorn


# utilities
six>=1.14
joblib==1.1.0
memoized-property
termcolor
textblob
xgboost==1.5.1
